// platform-api-payments/src/modules/payments/dtos/payment-intent-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class PaymentIntentResponseDto {
  @ApiProperty({
    description: 'The unique order ID generated by Razorpay.',
    example: 'order_ABC123XYZ',
  })
  gatewayOrderId: string;

  @ApiProperty({
    description: 'The amount to be paid, in the smallest currency unit.',
    example: 15000,
  })
  amount: number;

  @ApiProperty({
    description: 'The currency of the transaction.',
    example: 'INR',
  })
  currency: string;

  @ApiProperty({
    description: "The public Razorpay Key ID, required by the client's checkout SDK.",
    example: 'rzp_test_1234567890ABCD',
  })
  razorpayKeyId: string;

  @ApiProperty({
    description: 'The internal payment ID for our system.',
    example: 'f47ac10b-58cc-4372-a567-0e02b2c3d479',
  })
  paymentId: string;
}